+++ packages/frontend/theme/components/index.ts
@@ -1,3 +1,4 @@
 export * from './Menu';
 export * from './Button';
 export * from './Drawer';
+export * from './Alert';